Instrument with Sentry
* Breaks loadContext param in handleRequest in entry.server.ts

The loadContext variable is supposed to be passing the value created by
the function `getLoadContext` defined on line 106 of ./server/index.ts.
However, calling Sentry.init in ./app/entry.server.tsx results in the
loadContext variable being undefined in the handleRequest parameters.

This means that values from the server (such as the CSP nonce value for
script tags) can no longer be passed through to the remix server
application properly.